+++ Changelog.md
@@ -3,4 +3,5 @@
 
 ## 2.2.0
 
+- [#1425](https://github.com/wultra/enrollment-server/issues/1425) - Removed `Serializable` from JPA entities.
 - [#1775](https://github.com/wultra/enrollment-server/issues/1775) - Migrated to Spring Boot 4 and Jackson 3.